refactor(sanity): move reusable string-diff-related code to shared location
