+++ packages/sanity/src/core/form/inputs/common/diff/string/styles.ts
@@ -1,0 +1,19 @@
+import {css, type RuleSet} from 'styled-components'
+
+export function stringDiffContainerStyles(): RuleSet {
+  return css`
+    del {
+      opacity: 0.5;
+      text-decoration: line-through;
+
+      &::before {
+        text-decoration: line-through;
+        content: attr(data-text);
+      }
+    }
+
+    ins {
+      text-decoration: none;
+    }
+  `
+}
